requirements.txt
streamlit
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
pypdf2
python-dotenv
google-generativeai